agentic_rag_api/pyproject.toml
[project]
name = "agentic-rag-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.3.5",
    "fastapi>=0.121.3",
    "google-adk>=1.14.1",
    "google-generativeai>=0.8.5",
    "litellm>=1.80.7",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.4",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "httpx>=0.28.1",
    "pytest>=9.0.1",
]
